pyproject.toml
[project]
name = "apriltag-dataset"
version = "0.1.0"
description = "AprilTag image dataset with ground-truth detections"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pupil-apriltags>=1.0.4",
    "Pillow>=10.0",
    "imagehash>=4.3",
    "yt-dlp>=2024.0",
    "gdown>=5.0",
    "opencv-python-headless>=4.8",
]

[project.scripts]
apriltag-dataset = "apriltag_dataset.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/apriltag_dataset"]
